{#101} Refactoring the classes and interfaces. Adding the taskList bean to the main class for Test.
