+++ genie-core/src/main/java/com/netflix/genie/core/jobs/workflow/impl/CommandTask.java
@@ -1,7 +1,27 @@
-package com.netflix.genie.core.jobs;
+/*
+ *
+ *  Copyright 2015 Netflix, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ *
+ */
+package com.netflix.genie.core.jobs.workflow.impl;
 
 import com.netflix.genie.common.exceptions.GenieException;
 import com.netflix.genie.common.exceptions.GenieServerException;
+import com.netflix.genie.core.jobs.JobExecutionEnvironment;
+import com.netflix.genie.core.jobs.workflow.Context;
+import com.netflix.genie.core.jobs.workflow.WorkflowTask;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 
@@ -14,6 +34,7 @@ import java.nio.file.Path;
  * Implementation of the workflow task for processing command information.
  *
  * @author amsharma
+ * @since 3.0.0
  */
 @Slf4j
 public class CommandTask extends GenieBaseTask implements WorkflowTask {
