+++ genie-core/src/main/java/com/netflix/genie/core/jobs/workflow/impl/GenieBaseTask.java
@@ -1,4 +1,21 @@
-package com.netflix.genie.core.jobs;
+/*
+ *
+ *  Copyright 2015 Netflix, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ *
+ */
+package com.netflix.genie.core.jobs.workflow.impl;
 
 import com.google.common.collect.Lists;
 import com.netflix.genie.common.exceptions.GenieException;
@@ -17,6 +34,7 @@ import java.util.List;
  * helper methods that all classes can use.
  *
  * @author amsharma
+ * @since 3.0.0
  */
 public abstract class GenieBaseTask {
