+++ genie-core/src/test/java/com/netflix/genie/GenieCoreTestApplication.java
@@ -19,6 +19,13 @@ package com.netflix.genie;
 
 import com.github.springtestdbunit.bean.DatabaseConfigBean;
 import com.github.springtestdbunit.bean.DatabaseDataSourceConnectionFactoryBean;
+import com.netflix.genie.core.jobs.workflow.impl.ApplicationTask;
+import com.netflix.genie.core.jobs.workflow.impl.ClusterTask;
+import com.netflix.genie.core.jobs.workflow.impl.CommandTask;
+import com.netflix.genie.core.jobs.workflow.impl.IntialSetupTask;
+import com.netflix.genie.core.jobs.workflow.impl.JobKickoffTask;
+import com.netflix.genie.core.jobs.workflow.impl.JobTask;
+import com.netflix.genie.core.jobs.workflow.WorkflowTask;
 import org.dbunit.ext.hsqldb.HsqldbDataTypeFactory;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -29,6 +36,8 @@ import org.springframework.validation.beanvalidation.MethodValidationPostProcess
 
 import javax.sql.DataSource;
 import javax.validation.Validator;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Spring configuration class for integration tests.
@@ -84,5 +93,23 @@ public class  GenieCoreTestApplication {
                 = new DatabaseDataSourceConnectionFactoryBean(dataSource);
         dbConnection.setDatabaseConfig(dbUnitDatabaseConfig());
         return dbConnection;
+    }
+
+    /**
+     * Setup a bean to provide the list of impl in the workflow.
+     *
+     * @return List of workflow impl.
+     */
+    @Bean
+    // TODO Maybe return an empty list for testing? How is this used in testing.
+    public List<WorkflowTask> taskList() {
+        final List<WorkflowTask> taskList = new ArrayList<>();
+        taskList.add(new IntialSetupTask());
+        taskList.add(new ApplicationTask());
+        taskList.add(new CommandTask());
+        taskList.add(new ClusterTask());
+        taskList.add(new JobTask());
+        taskList.add(new JobKickoffTask());
+        return taskList;
     }
 }
